Added pipelines for claude 3.5 haiku and sonnet v2. Also fixed pipeline for sonnet v1, plus some refactoring
