+++ backend/open-webui-pipelines/.gitignore
@@ -7,8 +7,10 @@ pipelines/*
 !pipelines/.gitignore
 !pipelines/azure_openai_GPT35Turbo_pipeline.py
 !pipelines/azure_openai_GPT4Omni_pipeline.py
+!pipelines/bedrock_claude_haiku35_pipeline.py
 !pipelines/bedrock_claude_instant_pipeline.py
 !pipelines/bedrock_claude_sonnet35_pipeline.py
+!pipelines/bedrock_claude_sonnet35_v2_pipeline.py
 !pipelines/bedrock_llama32_11b_pipeline.py
 
 .DS_Store
